requirements.txt
streamlit
streamlit-chat
langchain
openai
wikipedia
chromadb
# tiktoken
python-dotenv
pyodbc
pandas
matplotlib
plotly